+++ examples/client_electron/renderer/index.js
@@ -10,11 +10,17 @@ function setContent (content) {
 
 document.getElementById('connect').addEventListener('click', () => {
   setContent('connecting...')
+  const authType = document.getElementById('type')
+
   const data = {
     host: document.getElementById('host').value,
     port: parseInt(document.getElementById('port').value),
     username: document.getElementById('username').value,
     password: document.getElementById('password').value === '' ? undefined : document.getElementById('password').value
+  }
+  if (authType.value === 'Microsoft') {
+    data.auth = 'microsoft'
+    delete data.password
   }
   ipcRenderer.send('connect', data)
 })
@@ -32,6 +38,11 @@ document.getElementById('chat').addEventListener('keyup', function onEvent (e) {
 document.getElementById('send').addEventListener('click', () => {
   chat()
 })
+
+window.onAuthTypeChange = function () {
+  const authType = document.getElementById('type')
+  console.log('set auth type to', authType)
+}
 
 ipcRenderer.on('content', (event, content) => {
   setContent(content)