+++ src/index.d.ts
@@ -49,6 +49,8 @@ declare module 'minecraft-protocol' {
 		stream?: Stream
 		connect?: (client: Client) => void
 		agent?: Agent
+		auth?: 'mojang' | 'microsoft'
+		id?: number
 	}
 
 	export class Server extends EventEmitter {
@@ -102,7 +104,7 @@ declare module 'minecraft-protocol' {
 	}
 
 	interface ClientsMap {
-		[key: string]: Client
+		[key: number]: Client
 	}
 
 	export interface PingOptions {